refactor(trajectory_validator): make traffic light compliance checker library (#3009)
* make separate traffic_light_compliance_checker library

- move tl compliance check code from trajecroy_validator filters to a separate package autoware_traffic_light_compliance_checker
- update trajectory_validator dependencies and remove obsolete filter files


* implement tl status tracker in traffic_light_compliance_checker library

- add new class TrafficLightStatusTracker to compliance checker library
- move signal history logic from traffic_light_filter to TrafficLightStatusTracker


* move amber_rejection_history logic to compliance checker library


* fix formatting


* revert unintended change


* add maintainer


---------
